requirements.txt
mcts>=1.0.4
Pillow>=8.1.0
